integration tests refactor cont
