+++ setup.cfg
@@ -56,7 +56,7 @@ max-line-length = 100
 
 [isort]
 known_first_party=pangeo_forge_recipes
-known_third_party=aiohttp,apache_beam,cftime,click,conftest,dask,fsspec,kerchunk,numpy,pandas,pytest,pytest_lazyfixture,s3fs,setuptools,ujson,xarray,zarr
+known_third_party=aiohttp,apache_beam,cftime,click,conftest,dask,fsspec,kerchunk,numpy,pandas,pytest,pytest_lazyfixture,s3fs,setuptools,xarray,zarr
 multi_line_output=3
 include_trailing_comma=True
 force_grid_wrap=0
